DATA/VERCEL/requirements.txt
flask==3.0.2
werkzeug==3.0.1
python-dotenv==1.0.1
requests==2.31.0
flask-cors==4.0.0
